dialogs added to main fragment when deleting, tikbumactivity almost done, sound problems, adding points not done, some changes

+++ PV239_game_trio/app/src/main/java/com/example/pv239_game_trio/frontend/main/start/MainFragment.kt
@@ -1,5 +1,7 @@
 package com.example.pv239_game_trio.frontend.main.start
 
+import android.app.AlertDialog
+import android.content.DialogInterface
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
@@ -13,6 +15,8 @@ import androidx.recyclerview.widget.RecyclerView
 import androidx.room.Room
 import com.example.pv239_game_trio.R
 import com.example.pv239_game_trio.backend.AppDB
+import com.example.pv239_game_trio.backend.dto.PlayerDTO
+import com.example.pv239_game_trio.backend.entities.PlayerEntity
 import com.example.pv239_game_trio.frontend.main.ChooseGameActivity
 
 
@@ -35,6 +39,7 @@ class MainFragment : Fragment() {
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         val view : View = inflater.inflate(R.layout.activity_main_fragment, container, false)
         val db = Room.databaseBuilder<AppDB>(context!!.applicationContext, AppDB :: class.java, "GameTrioDB").build()
+        createPlayers(db)
 
         startActivityButton = view.findViewById(R.id.button_start)
         addPlayerButton = view.findViewById(R.id.button_add)
@@ -48,6 +53,10 @@ class MainFragment : Fragment() {
         infoText = view.findViewById(R.id.textView_info)
         players = view.findViewById(R.id.textView_players)
 
+
+
+        players.setText(showPlayers(db))
+
         startActivityButton.setOnClickListener(View.OnClickListener {
             Log.d(TAG,"button START was pressed")
             openActivityMain()
@@ -57,10 +66,10 @@ class MainFragment : Fragment() {
         addPlayerButton.setOnClickListener(View.OnClickListener {
             Log.d(TAG,"button ADD PLAYER was pressed")
             val fragment = AddPlayerFragment()
-            //replaceFragment(fragment)
+            replaceFragment(fragment)
             //val activity = MainActivity()
 
-            //MainActivity().setViewPager(0)
+            MainActivity().setViewPager(0)
         })
 
         //TODO
@@ -78,18 +87,43 @@ class MainFragment : Fragment() {
 
         resetPointsButton.setOnClickListener(View.OnClickListener {
             Log.d(TAG,"button RESET POINTS was pressed")
-            Thread{
-                //db.playerDAO().resetPointAllPlayers()
-                Log.d(TAG,"resetPointAllPlayers() DONE")
-            }.start()
+
+            val builder : AlertDialog.Builder = AlertDialog.Builder(context)
+            builder.setMessage("Are you sure you want to reset all points?")
+            builder.setPositiveButton("Yes"){dialog, which ->
+                    Thread{
+                        db.playerDAO().resetPointAllPlayers()
+                        Log.d(TAG,"resetPointAllPlayers() DONE")
+                    }.start()
+                }
+            builder.setNegativeButton("No"){dialog, which ->
+                Log.d(TAG,"resetPointAllPlayers() canceled by user")
+
+            }
+            builder.show()
+
+
+
         })
 
         restartGameButton.setOnClickListener(View.OnClickListener {
             Log.d(TAG,"button RESTART GAME was pressed")
-            Thread{
-                //db.playerDAO().deleteAllPlayers()
-                Log.d(TAG,"deleteAllPlayers() DONE")
-            }.start()
+
+
+
+            val builder : AlertDialog.Builder = AlertDialog.Builder(context)
+            builder.setMessage("Are you sure you want to restart game?")
+            builder.setPositiveButton("Yes"){dialog, which ->
+                Thread{
+                    db.playerDAO().deleteAllPlayers()
+                    Log.d(TAG,"deleteAllPlayers() DONE")
+                }.start()
+            }
+            builder.setNegativeButton("No"){dialog, which ->
+                Log.d(TAG,"deleteAllPlayers() canceled by user")
+
+            }
+            builder.show()
         })
 
         return view
@@ -105,6 +139,67 @@ class MainFragment : Fragment() {
         transaction.replace(R.id.container, someFragment)
         transaction.addToBackStack(null)
         transaction.commit()
+    }
+
+
+    private fun showPlayers(db: AppDB) : String{
+        var actualPlayers = ""
+
+        Thread{
+            var playersInDB = db.playerDAO().showAllPlayers()
+
+            for (playerDb in playersInDB){
+                var playerDTO = PlayerDTO()
+                playerDTO.name = playerDb.name
+                playerDTO.points = playerDb.points
+                playerDTO.team = playerDb.team
+                actualPlayers = actualPlayers + playerDTO.toString() + "/n"
+            }
+        }.start()
+
+        return actualPlayers
+    }
+
+
+    private fun createPlayers(db: AppDB){
+        Thread {
+            var player1 = PlayerEntity()
+            player1.id = 0
+            player1.name = "Peter"
+            player1.team = 1
+            player1.points = 0
+
+
+            var player2 = PlayerEntity()
+            player2.id = 1
+            player2.name = "Marcel"
+            player2.team = 1
+            player2.points = 0
+
+
+            var player3 = PlayerEntity()
+            player3.id = 2
+            player3.name = "Jitka"
+            player3.team = 0
+            player3.points = 0
+
+
+            var player4 = PlayerEntity()
+            player4.id = 3
+            player4.name = "Sara"
+            player4.team = 0
+            player4.points = 0
+
+
+
+            db.playerDAO().create(player1)
+            db.playerDAO().create(player2)
+            db.playerDAO().create(player3)
+            db.playerDAO().create(player4)
+
+            Log.d(TAG,"4 players created")
+
+        }.start()
     }
